Update Ion  & Ion-gui Toc interface version to 60000  & addon version to WoD-Beta-Update Updated UpdateIconIndex to reflect change to GetMacroIcons which now returns texture file data instead of path. Updated any place that set icon texture by path to also check and set via file data.
Added list of current mounts spellIDs referenced to the ActionID returned by GetCursorInfo when a mount is held. The ActionID does not match the MountID used in C_MountJournal.GetMountInfo, so the mount table is paired to a generated table of the items in the MountJournal via the spell IDs.  Note the items in the MountJournal index is sorted alphabetically and order changes based on player's localization, which is why the tables are not hard coded.  Until a better way to map CursorInfo's ActionID to the MountJournal index, the mount table will need to be manually updated any time a new mount is added to the game.